pyproject.toml
[project]
name = "zotero-cli"
version = "2.0.0-dev"
dependencies = [
    "requests>=2.31.0",
    "rich>=13.7.0",
    "toml>=0.10.2",
    "tenacity>=8.2.3",
    "pytest>=7.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (F), pycodestyle (E, W), isort (I)
select = ["F", "E", "W", "I"]
ignore = [
    "E501", # Line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["zotero_cli"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false # We'll tighten this gradually
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that interact with real APIs",
]